requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Migrate==4.0.4
python-dotenv==1.0.0
Werkzeug==2.3.7
PyPDF2==3.0.1
langchain==0.0.267
langchain-openai==0.0.2
pdfplumber==0.10.2
gunicorn==21.2.0
